+++ src/main/java/codesquad/Main.java
@@ -1,7 +1,7 @@
 package codesquad;
 
 import codesquad.http.MIME;
-import codesquad.server.Server;
+import codesquad.http.Server;
 import codesquad.server.handlers.CreateUserHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -17,7 +17,6 @@ public class Main {
     public static void main(String[] args) {
         MIME.init();
         Server server = new Server(PORT, THREAD_POOL_SIZE);
-
         server.get("/create", CreateUserHandler::createUser);
         server.staticFiles("/", "/static");
